+++ services/boilerplate/hf_service_handler.py
@@ -279,6 +279,25 @@ class ForecastingHuggingFaceHandler(ForecastingServiceHandler, HuggingFaceHandle
         return self
 
 
+class TinyTimeMixerForecastingHandler(ForecastingHuggingFaceHandler):
+    """Service handler for the tiny time mixer model"""
+
+    def _get_config_kwargs(
+        self,
+        parameters: Optional[ForecastingParameters] = None,
+        preprocessor: Optional[TimeSeriesPreprocessor] = None,
+    ) -> Dict[str, Any]:
+        config_kwargs = {
+            "num_input_channels": preprocessor.num_input_channels,
+            "prediction_filter_length": parameters.prediction_length,
+            "exogenous_channel_indices": preprocessor.exogenous_channel_indices,
+            "prediction_channel_indices": preprocessor.prediction_channel_indices,
+        }
+        return config_kwargs
+
+
+
+
 def register_config(model_type: str, model_config_name: str, module_path: str) -> None:
     """Register a configuration for a particular model architecture
