+++ services/finetuning/tsfmfinetuning/.gitignore
@@ -7,4 +7,3 @@ hfutil.py
 inference_payloads.py
 service_handler.py
 tsfm_config.py
-tsfm_service_handler.py
